backend/requirements.txt
django
django-rest-framework
requests
django-cors-headers